:recycle: Refactor!
