+++ src/routes/room/[room=uuid]/+page.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
+	import ViewSchedules from './ViewSchedules.svelte';
+
 	import * as Tabs from '$lib/components/ui/tabs';
-	import ScheduleDisplay from './ScheduleDisplay.svelte';
 	import InfoInput from '$lib/InfoInput.svelte';
 
 	import { page } from '$app/stores';
@@ -8,23 +9,14 @@
 	import { onMount } from 'svelte';
 	import { writable } from 'svelte/store';
 
-	import type { VirtualSchedule, Classes, Class } from '$lib/InfoInput.d';
+	import type { VirtualSchedule, Classes, Class, Schedule } from '$lib/InfoInput.d';
 	import type { Writable } from 'svelte/store';
 	import memoize from 'lodash-es/memoize';
-	import isEqual from 'lodash-es/isEqual';
 	import Toasts from '$lib/Toasts.svelte';
 	import { addToast } from '$lib/toasts';
-	// import type { PageServerData } from './$types';
 
 	let onlyMatching: boolean;
-	const PERIODS: (keyof VirtualSchedule)[] = ['1a', '2a', '3a', '4a', '1b', '2b', '3b', '4b'];
-	function matches(a: VirtualSchedule, b: VirtualSchedule) {
-		return PERIODS.some((x) => a[x] === b[x]);
-	}
-	type Schedule = VirtualSchedule & {
-		room: string;
-		student: string;
-	};
+
 	export let data;
 	const { supabase } = data;
 	const schedules: Writable<Schedule[]> = writable(data.data);
@@ -275,73 +267,7 @@
 		</Tabs.List>
 	{/if}
 	<Tabs.Content value="schedules">
-		<div class="flex flex-wrap justify-evenly">
-			{#each $schedules as schedule}
-				<!-- You is guaranteed to !== null -->
-				<!-- I'm not sure how to express TypeScript -->
-				<!-- Within Svelte code -->
-				{#if you === 'tentative'}
-					<button
-						class="btn my-3 h-fit w-fit border border-base-300 bg-base-100 shadow-xl rounded-box p-4 text-xl font-medium"
-						on:click={() => {
-							you = { name: schedule.student, schedule };
-							window.localStorage.setItem($page.params.room, JSON.stringify(you));
-						}}
-					>
-						{schedule.student}
-					</button>
-				{:else if you === null}
-					<p>Please input who you are first</p>
-				{:else if (onlyMatching && matches(you.schedule, schedule)) || !onlyMatching}
-					<div
-						class="my-3 collapse h-fit w-fit collapse-plus border border-base-300 bg-base-100 shadow-xl rounded-box"
-					>
-						<input type="checkbox" />
-						<div class="collapse-title text-xl font-medium">
-							{schedule.student}'s schedule {isEqual({ name: schedule.student, schedule }, you)
-								? '(you)'
-								: ''}
-							<!-- TODO: Show if in common -->
-						</div>
-						<div class="collapse-content hidden">
-							<div class="overflow-x-auto">
-								<!-- If statement to appease type checker -->
-								{#if you != null}
-									<ScheduleDisplay them={schedule} you={you.schedule} {getClass} />
-								{/if}
-							</div>
-						</div>
-					</div>
-				{/if}
-			{:else}
-				<div class="alert alert-warning shadow-lg mx-auto w-fit">
-					<div>
-						<svg
-							xmlns="http://www.w3.org/2000/svg"
-							class="stroke-current flex-shrink-0 h-6 w-6"
-							fill="none"
-							viewBox="0 0 24 24"
-							><path
-								stroke-linecap="round"
-								stroke-linejoin="round"
-								stroke-width="2"
-								d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
-							/></svg
-						>
-						<span
-							>No schedules found. <button
-								class="link link-primary"
-								on:click={() => {
-									navigator.clipboard.writeText(window.location.href).then(() => {
-										addToast('Room URL copied to clipboard', 'success');
-									});
-								}}>Invite people!</button
-							></span
-						>
-					</div>
-				</div>
-			{/each}
-		</div>
+		<ViewSchedules {schedules} {you} room={$page.params.room} {getClass} {onlyMatching} />
 	</Tabs.Content>
 	<Tabs.Content value="password">Change your password here.</Tabs.Content>
 </Tabs.Root>
@@ -357,9 +283,6 @@
 </span>
 
 <style>
-	input[type='checkbox']:checked ~ .collapse-content {
-		display: block;
-	}
 	.center-horizontal {
 		/*shift 50% of the page*/
 		left: 50%;
